restructures the frameworks to simplify them a bit.
